move and rename some classes
* move CommandLineParser + CommandLineParams to separate file
* rename CommandLineArgsParserApacheCommonsImpl
