+++ build.gradle.kts
@@ -138,6 +138,8 @@ val jacocoTestCoverageVerification =
                 excludes = listOf(
                     "*.DefaultImpls", "*.Companion", "*.Factory", "*.1",
                     "org.example.subtitles.modification.SubtitlesTransformer",
+                    "org.example.subtitles.cli.CommandLineArgsParserApacheCommonsImpl",
+                    "org.example.subtitles.cli.CommandLineArgsParser",
                     "org.example.subtitles.cli.SimpleCliKt",
                     "org.example.subtitles.cli.TimedSubtitlePrinter"
                 )
